[Fix] : 전체 조회, 페이스메이커 조회시 썸네일 조회 이슈 수정
