+++ src/main/java/com/mogak/spring/converter/PostConverter.java
@@ -100,7 +100,8 @@ public class PostConverter {
                 .userJob(post.getUser().getJob().getName())
                 .contents(post.getContents())
                 .imgUrls(post.getPostImgs().stream()
-                        .map(m -> m.getImgUrl())
+                        .filter(img -> img.getImgUrl() != post.getPostThumbnailUrl())
+                        .map(PostImg::getImgUrl)
                         .collect(Collectors.toList()))
                 .commentCnt(post.getCommentCnt())
                 .likeCnt(post.getLikeCnt())
