--- src/main/java/com/mogak/spring/service/PostImgService.java
@@ -1,13 +1,0 @@
-package com.mogak.spring.service;
-
-import com.mogak.spring.domain.post.Post;
-import com.mogak.spring.domain.post.PostImg;
-
-import java.util.List;
-
-public interface PostImgService {
-
-    List<String> findUrlByPost(Long postId);
-    List<String> findNotThumbnailImg(Post post);
-    List<PostImg> findAllByPost(Post post);
-}
